requirements.txt
grpcio
grpcio-tools
paho-mqtt